requirements.txt
pytrends==4.9.2
flask>=3.0
pandas==2.2.2
requests>=2.31.0
python-dotenv>=1.0.0
